✨ feat     : Manage pagination state according to each page
Utilized state management instead of query parameters for the 'My Page' component, which comprises two sections: detailed posts and a paginated list. Meanwhile, for the list page where pagination is prominent, introduced management of the 'currentPage' state through query parameters. This optimization aims to reduce unnecessary rendering.
