+++ src/app/mypage/inquiry/page.tsx
@@ -1,21 +1,13 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { InquiryPost, InquiryPostList } from '@/types/apis/userInquiry';
+import { InquiryPost } from '@/types/apis/userInquiry';
 import InquiryDetail from '../_components/Inquiry/InquiryDetail/InquiryDetail';
 import InquiryList from '../_components/Inquiry/InquiryList/InquiryList';
 import * as S from './inquiry.style';
 
-export default function Page({
-  searchParams,
-}: {
-  searchParams: {
-    [key: string]: string | string[] | undefined;
-  };
-}) {
+export default function Page() {
   const [inquiryData, setInquiryData] = useState<InquiryPost | null>(null);
-  const [inquiryListItemData, setInquiryListItemData] =
-    useState<InquiryPostList | null>(null);
 
   useEffect(() => {
     //TODO react-query로 변경 및 api 분리
@@ -27,19 +19,6 @@ export default function Page({
 
     fetchInquiry();
   }, []);
-
-  useEffect(() => {
-    //TODO react-query로 변경 및 api 분리
-    const fetchInquiryList = async () => {
-      const inquiryListItemData = await fetch(
-        `/mock/inquiryList${searchParams.page}.json`,
-      );
-      const result = await inquiryListItemData.json();
-      setInquiryListItemData(result.data);
-    };
-
-    fetchInquiryList();
-  }, [searchParams.page]);
 
   return (
     <S.InquiryContainer>
@@ -66,9 +45,7 @@ export default function Page({
             )}
           </S.InquiryWrapper>
 
-          {inquiryListItemData && (
-            <InquiryList inquiryListItemData={inquiryListItemData} />
-          )}
+          <InquiryList />
         </S.InquiryBox>
       )}
     </S.InquiryContainer>
